utils cleanup
