+++ src/js/techreport/utils/data.js
@@ -1,21 +1,3 @@
-// Get the color associated with the techology
-// Based on the order of the technologies
-// To ensure the color is consistent per query
-const getAppColor = (tech, technologies, colors) => {
-  const sortedTechs = [...technologies].sort((a, b) => a.toLowerCase().localeCompare(b.toLowerCase()));
-  const techIndex = sortedTechs.indexOf(tech);
-
-  // Return custom colors if configured
-  if(colors.overrides && colors.overrides[tech]) {
-    return colors.overrides[tech];
-  }
-
-  // Otherwise reutrn based on alphabetic position
-  if(techIndex < colors.app.length) {
-    return colors.app[techIndex];
-  }
-}
-
 const parseVitalsData = (metric, date) => {
   return metric.map(submetric => {
     return {
@@ -92,10 +74,9 @@ const parsePageWeightData = (metric, date) => {
   });
 }
 
-export const Utils = {
-  getAppColor,
+export const DataUtils = {
   parseVitalsData,
   parseLighthouseData,
   parseAdoptionData,
   parsePageWeightData,
-}
+};
